+++ src/records/EntryCountRecord.php
@@ -8,22 +8,15 @@ namespace putyourlightson\entrycount\records;
 use craft\db\ActiveRecord;
 
 /**
- * EntryCountRecord
- *
- * @property int         $id                         ID
- * @property int         $entryId                    Entry ID
- * @property int         $count                      Count
+ * @property int $id
+ * @property int $entryId
+ * @property int $count
  */
 class EntryCountRecord extends ActiveRecord
 {
-    // Public Static Methods
-    // =========================================================================
-
-     /**
-     * @inheritdoc
-     *
-     * @return string the table name
-     */
+    /**
+    * @inheritdoc
+    */
     public static function tableName(): string
     {
         return '{{%entrycount}}';