+++ src/translations/hu/entry-count.php
@@ -5,7 +5,7 @@
 return [
     'Count' => 'Számláló',
     'Entry count reset.' => 'Számláló visszaállítása',
-    'Ignore logged in users' => 'Bejelentkezett felhasználók figyelmen kívül hagyása', 
+    'Ignore logged in users' => 'Bejelentkezett felhasználók figyelmen kívül hagyása',
     'Only increment counts if users are not logged in' => 'Csak abban az esetben növekedik a megtekintések száma, ha a felhasználó nincs bejelentkezve',
     'Ignore IP addresses' => 'IP címek figyelmen kívül hagyása',
     'Add one IP address per line to ignore them when incrementing counts' => 'A hozzáadott IP címekről (mindegyik új sorban) érkező kérések esetén nem fog növekedni a megtekintések száma',
@@ -14,5 +14,5 @@ return [
     'URL' => 'URL',
     'First Count' => 'Első megtekintés',
     'Last Count' => 'Utolsó megtekintés',
-    'No entries counted' => 'Még nincsenek megtekintések'
+    'No entries counted' => 'Még nincsenek megtekintések',
 ];